+++ projects/js-packages/i18n-loader-webpack-plugin/tests/fixtures/ignore-modules/tests.js
@@ -1,5 +1,4 @@
-/* global wpI18n */
-const { default: md5 } = require( 'md5-es' );
+/* global jpI18nLoader, wpI18n */
 
 describe( 'Test ignoreModules as', () => {
 	beforeEach( () => {
@@ -10,16 +9,12 @@ describe( 'Test ignoreModules as', () => {
 	test.each( whats )( 'a %s', async what => {
 		const main = require( `./dist/${ what }/main.js` );
 		const main2 = require( `./dist/${ what }/main2.js` );
-		const hash = md5.hash( `dist/${ what }/hasI18n2.js` );
 
 		expect( await main.noI18n() ).toEqual( 'No i18n here' );
 		expect( await main.hasI18n() ).toEqual( 'This is translated' );
 
 		expect( await main2.hasI18n() ).toEqual( 'This is translated' );
-		fetch.expectUrl(
-			`http://test.example.com/wp-content/languages/plugins/ignore-modules-en_piglatin-${ hash }.json`,
-			require( './en_piglatin.json' )
-		);
+		jpI18nLoader.expectI18n( `dist/${ what }/hasI18n2.js`, require( './en_piglatin.json' ) );
 		expect( await main2.hasI18n2() ).toEqual( 'is-Thay is-way anslated-tray oo-tay' );
 	} );
 } );