+++ projects/packages/assets/tests/php/test-assets-files/i18n-map.php
@@ -3,7 +3,10 @@ return array(
 	'domain' => 'target',
 	'type' => 'plugins',
 	'packages' => array(
-		'foo' => '1.2.3',
+		'foo' => array(
+			'path' => 'path/to/foo',
+			'ver' => '1.2.3',
+		),
 		'bar' => '4.5.6',
 	),
 );