+++ projects/packages/my-jetpack/_inc/state/store.js
@@ -1,6 +1,7 @@
 /**
  * Internal dependencies
  */
+import actions from './actions';
 import reducer from './reducers';
 import selectors from './selectors';
 import storeHolder from './store-holder';
@@ -14,8 +15,9 @@ const STORE_ID = 'my-jetpack';
  */
 function initStore() {
 	storeHolder.mayBeInit( STORE_ID, {
+		__experimentalUseThunks: true, // never stop experiment :sweat_smile:
 		reducer,
-		actions: {},
+		actions,
 		selectors,
 		resolvers,
 		controls,