+++ projects/packages/my-jetpack/src/class-products.php
@@ -63,6 +63,20 @@ class Products {
 			'title'      => 'The requested product data',
 			'type'       => 'object',
 			'properties' => array(
+				'product'     => array(
+					'description'       => __( 'Product slug', 'jetpack-my-jetpack' ),
+					'type'              => 'string',
+					'enum'              => __CLASS__ . '::get_product_names',
+					'required'          => false,
+					'validate_callback' => __CLASS__ . '::check_product_argument',
+				),
+				'action'      => array(
+					'description'       => __( 'Production action to execute', 'jetpack-my-jetpack' ),
+					'type'              => 'string',
+					'enum'              => array( 'activate', 'deactivate' ),
+					'required'          => false,
+					'validate_callback' => __CLASS__ . '::check_product_argument',
+				),
 				'slug'        => array(
 					'title' => 'The product slug',
 					'type'  => 'string',
@@ -94,12 +108,7 @@ class Products {
 	 * @return array Object with infromation about the product.
 	 */
 	public static function get_anti_spam_data() {
-		return array(
-			'slug'        => 'anti-spam',
-			'description' => __( 'Stop comment and form spam', 'jetpack-my-jetpack' ),
-			'name'        => __( 'Anti-spam', 'jetpack-my-jetpack' ),
-			'status'      => 'inactive',
-		);
+		return Products\Anti_Spam::get_info();
 	}
 
 	/**
@@ -108,12 +117,7 @@ class Products {
 	 * @return array Object with infromation about the product.
 	 */
 	public static function get_backup_data() {
-		return array(
-			'slug'        => 'backup',
-			'description' => __( 'Save every change', 'jetpack-my-jetpack' ),
-			'name'        => __( 'Backup', 'jetpack-my-jetpack' ),
-			'status'      => 'active',
-		);
+		return Products\Backup::get_info();
 	}
 
 	/**
@@ -131,12 +135,7 @@ class Products {
 	 * @return array Object with infromation about the product.
 	 */
 	public static function get_crm_data() {
-		return array(
-			'slug'        => 'crm',
-			'description' => __( 'Connect with your people', 'jetpack-my-jetpack' ),
-			'name'        => __( 'CRM', 'jetpack-my-jetpack' ),
-			'status'      => 'plugin_absent',
-		);
+		return Products\Crm::get_info();
 	}
 
 	/**
@@ -159,12 +158,7 @@ class Products {
 	 * @return array Object with infromation about the product.
 	 */
 	public static function get_scan_data() {
-		return array(
-			'slug'        => 'scan',
-			'description' => __( 'Stay one step ahead of threats', 'jetpack-my-jetpack' ),
-			'name'        => __( 'Scan', 'jetpack-my-jetpack' ),
-			'status'      => 'plugin_absent',
-		);
+		return Products\Scan::get_info();
 	}
 
 	/**
@@ -173,12 +167,7 @@ class Products {
 	 * @return array Object with infromation about the product.
 	 */
 	public static function get_search_data() {
-		return array(
-			'slug'        => 'search',
-			'description' => __( 'Help them find what they need', 'jetpack-my-jetpack' ),
-			'name'        => __( 'Search', 'jetpack-my-jetpack' ),
-			'status'      => 'plugin_absent',
-		);
+		return Products\Search::get_info();
 	}
 
 	/**
@@ -187,11 +176,6 @@ class Products {
 	 * @return array Object with infromation about the product.
 	 */
 	public static function get_videopress_data() {
-		return array(
-			'slug'        => 'videopress',
-			'description' => __( 'High quality, ad-free video', 'jetpack-my-jetpack' ),
-			'name'        => __( 'VideoPress', 'jetpack-my-jetpack' ),
-			'status'      => 'active',
-		);
+		return Products\Videopress::get_info();
 	}
 }