+++ projects/plugins/boost/app/assets/src/js/utils/analytics.ts
@@ -5,6 +5,7 @@ export function recordBoostEvent(
 ): void {
 	// eslint-disable-next-line camelcase
 	if ( ! ( 'boost_version' in eventProp ) && 'version' in Jetpack_Boost ) {
+		// eslint-disable-next-line camelcase
 		eventProp.boost_version = Jetpack_Boost.version;
 	}