+++ projects/plugins/boost/app/lib/class-connection.php
@@ -9,6 +9,7 @@
 
 namespace Automattic\Jetpack_Boost\Lib;
 
+use Automattic\Jetpack\Config as Jetpack_Config;
 use Automattic\Jetpack\Connection\Manager;
 use Automattic\Jetpack\Terms_Of_Service;
 
@@ -290,5 +291,19 @@ class Connection {
 	 */
 	public static function rest_authorization_required_code() {
 		return is_user_logged_in() ? 403 : 401;
+	}
+
+	public function ensure_connection() {
+		if ( ! apply_filters( 'jetpack_boost_connection_bypass', false ) ) {
+			$jetpack_config = new Jetpack_Config();
+			$jetpack_config->ensure(
+				'connection',
+				array(
+					'slug'     => 'jetpack-boost',
+					'name'     => 'Jetpack Boost',
+					'url_info' => '', // Optional, URL of the plugin.
+				)
+			);
+		}
 	}
 }