+++ tools/cli/commands/install.js
@@ -2,48 +2,53 @@
  * External dependencies
  */
 import chalk from 'chalk';
+import execa from 'execa';
+import fs from 'fs/promises';
 import Listr from 'listr';
+import VerboseRenderer from 'listr-verbose-renderer';
+import UpdateRenderer from 'listr-update-renderer';
 
 /**
  * Internal dependencies
  */
 import promptForProject from '../helpers/promptForProject.js';
-import installProjectTask from '../helpers/tasks/installProjectTask.js';
+import { needsPnpmInstall, getInstallArgs, projectDir } from '../helpers/install.js';
 import { allProjects } from '../helpers/projectHelpers.js';
-import { normalizeInstallArgv, normalizeProject } from '../helpers/normalizeArgv.js';
-import listrOpts from '../helpers/tasks/listrOpts.js';
+
+export const command = 'install [project...]';
+export const describe = 'Installs a monorepo project';
 
 /**
- * Installs a project.
+ * Options definition for the install subcommand.
  *
- * @param {object} argv - Passthrough of the argv to include project, root, etc options.
+ * @param {object} yargs - The Yargs dependency.
+ * @returns {object} Yargs with the install commands defined.
  */
-export async function install( argv ) {
-	argv = normalizeInstallArgv( argv );
-
-	let tasks = [];
-	if ( argv.all ) {
-		allProjects().forEach( item => {
-			argv.project = item;
-			tasks.push( installProjectTask( argv ) );
+export function builder( yargs ) {
+	return yargs
+		.positional( 'project', {
+			describe: 'Project in the form of type/name, e.g. plugins/jetpack',
+			type: 'string',
+		} )
+		.option( 'root', {
+			alias: 'r',
+			type: 'boolean',
+			description: 'Install the monorepo dependencies',
+		} )
+		.option( 'all', {
+			alias: 'a',
+			type: 'boolean',
+			description: 'Installs everything',
+		} )
+		.option( 'no-pnpm-install', {
+			type: 'boolean',
+			description: 'Skip execution of `pnpm install`.',
+		} )
+		.option( 'concurrency', {
+			type: 'number',
+			description: 'Maximum number of install tasks to run at once. Ignored with `--verbose`.',
+			default: Infinity,
 		} );
-		// Reset.
-		argv.root = true;
-		argv.project = '';
-	}
-
-	tasks = addRootInstallTask( argv, tasks );
-
-	argv.project ? tasks.push( installProjectTask( argv ) ) : null;
-
-	const opts = listrOpts( argv );
-
-	const installs = new Listr( tasks, opts );
-
-	installs.run().catch( err => {
-		console.error( err );
-		process.exit( err.exitCode || 1 );
-	} );
 }
 
 /**
@@ -51,68 +56,70 @@ export async function install( argv ) {
  *
  * @param {object} argv - The argv for the command line.
  */
-export async function installCli( argv ) {
-	argv = normalizeInstallArgv( argv );
-	argv = normalizeProject( argv );
-
-	if ( ! argv.root && ! argv.all ) {
-		argv = await promptForProject( argv );
-	}
-
-	if ( argv.project || argv.root || argv.all ) {
-		await install( argv );
-	} else {
-		console.error( chalk.red( 'You did not choose a valid project!' ) );
-	}
-}
-
-/**
- * Command definition for the install subcommand.
- *
- * @param {object} yargs - The Yargs dependency.
- * @returns {object} Yargs with the install commands defined.
- */
-export function installDefine( yargs ) {
-	yargs.command(
-		'install [project]',
-		'Installs a monorepo project',
-		yarg => {
-			yarg
-				.positional( 'project', {
-					describe: 'Project in the form of type/name, e.g. plugins/jetpack',
-					type: 'string',
-				} )
-				.option( 'root', {
-					alias: 'r',
-					type: 'boolean',
-					description: 'Install the monorepo dependencies',
-				} )
-				.option( 'all', {
-					alias: 'a',
-					type: 'boolean',
-					description: 'Installs everything',
-				} );
-		},
-		async argv => {
-			await installCli( argv );
-			if ( argv.v ) {
-				console.log( argv );
-			}
+export async function handler( argv ) {
+	if ( argv.project.length === 1 ) {
+		if ( argv.project[ 0 ].indexOf( '/' ) < 0 ) {
+			argv.type = argv.project[ 0 ];
+			argv.project = [];
 		}
-	);
+	}
 
-	return yargs;
-}
+	if ( argv.all ) {
+		argv.project = allProjects();
+	}
+	if ( argv.all || argv.root ) {
+		argv.project.unshift( 'monorepo' );
+	}
 
-/**
- * Conditionally add the root install task
- *
- * @param {object} argv - The argv object
- * @param {Array} tasks - The tasks array
- * @returns {Array} The tasks array
- */
-function addRootInstallTask( argv, tasks ) {
-	normalizeInstallArgv( argv );
-	argv.root ? tasks.push( installProjectTask( argv ) ) : null;
-	return tasks;
+	if ( argv.project.length === 0 ) {
+		argv.project = '';
+		argv = await promptForProject( argv );
+		argv.project = [ argv.project ];
+	}
+
+	const stdio = argv.v ? [ 'ignore', 'inherit', 'inherit' ] : [ 'ignore', 'ignore', 'ignore' ];
+	const tasks = [];
+	let didPnpm = false;
+
+	for ( const project of new Set( argv.project ) ) {
+		// Does the project even exist?
+		if (
+			( await fs.access( projectDir( project, 'composer.json' ) ).catch( () => false ) ) === false
+		) {
+			console.error( chalk.red( `Project ${ project } does not exist!` ) );
+			continue;
+		}
+
+		// Do we need pnpm for this project?
+		if ( argv.pnpmInstall !== false && ! didPnpm && ( await needsPnpmInstall( project ) ) ) {
+			didPnpm = true;
+			tasks.unshift( {
+				title: `Installing pnpm dependencies`,
+				task: async () =>
+					execa( 'pnpm', await getInstallArgs( 'monorepo', 'pnpm', argv ), {
+						cwd: process.cwd(),
+						stdio,
+					} ),
+			} );
+		}
+
+		// Composer install.
+		tasks.push( {
+			title: `Installing composer dependencies for ${ project }`,
+			task: async () =>
+				execa( 'composer', await getInstallArgs( project, 'composer', argv ), {
+					cwd: projectDir( project ),
+					stdio,
+				} ),
+		} );
+	}
+
+	const listr = new Listr( tasks, {
+		concurrent: argv.v ? false : argv.concurrency,
+		renderer: argv.v ? VerboseRenderer : UpdateRenderer,
+	} );
+	await listr.run().catch( err => {
+		console.error( err );
+		process.exit( err.exitCode || 1 );
+	} );
 }