+++ tools/cli/commands/test.js
@@ -2,6 +2,7 @@
  * External dependencies
  */
 import chalk from 'chalk';
+import execa from 'execa';
 import inquirer from 'inquirer';
 import child_process from 'child_process';
 import path from 'path';
@@ -12,7 +13,7 @@ import UpdateRenderer from 'listr-update-renderer';
 /**
  * Internal dependencies
  */
-import installProjectTask from '../helpers/tasks/installProjectTask.js';
+import { getInstallArgs, projectDir } from '../helpers/install.js';
 import promptForProject from '../helpers/promptForProject.js';
 import { readComposerJson } from '../helpers/json.js';
 import { allProjects } from '../helpers/projectHelpers.js';
@@ -137,9 +138,27 @@ async function getTestInstructions( argv ) {
  */
 async function runTest( argv ) {
 	const installer = new Listr(
-		[ installProjectTask( { project: argv.project, v: argv.verbose } ) ],
+		[
+			{
+				title: `Installing pnpm dependencies`,
+				task: async () =>
+					execa( 'pnpm', await getInstallArgs( 'monorepo', 'pnpm', argv ), {
+						cwd: process.cwd(),
+						stdio: argv.v ? 'inherit' : 'ignore',
+					} ),
+			},
+			{
+				title: `Installing composer dependencies`,
+				task: async () =>
+					execa( 'composer', await getInstallArgs( argv.project, 'composer', argv ), {
+						cwd: projectDir( argv.project ),
+						stdio: argv.v ? 'inherit' : 'ignore',
+					} ),
+			},
+		],
 		{
-			renderer: argv.verbose ? VerboseRenderer : UpdateRenderer,
+			concurrent: ! argv.v,
+			renderer: argv.v ? VerboseRenderer : UpdateRenderer,
 		}
 	);
 	await installer.run();