+++ projects/js-packages/api/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## 0.8.3 - 2022-01-25
+### Changed
+- Updated package dependencies.
+
 ## 0.8.2 - 2022-01-18
 ### Changed
 - General: update required node version to v16.13.2