+++ projects/js-packages/webpack-config/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.1 - 2022-01-27
+### Changed
+- Updated package dependencies.
+
+## 1.1.0 - 2022-01-25
+### Added
+- Add missing dev dependency on `nyc` for code coverage.
+
+### Changed
+- Updated package dependencies. Major version bump for i18n-loader-webpack-plugin.
+
 ## 1.0.2 - 2022-01-18
 ### Changed
 - General: update required node version to v16.13.2