+++ projects/packages/error/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.13] - 2022-01-25
+### Changed
+- Updated package dependencies.
+
 ## [1.3.12] - 2022-01-04
 ### Changed
 - Switch to pcov for code coverage.
@@ -92,6 +96,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Introduce a jetpack-error package
 
+[1.3.13]: https://github.com/Automattic/jetpack-error/compare/v1.3.12...v1.3.13
 [1.3.12]: https://github.com/Automattic/jetpack-error/compare/v1.3.11...v1.3.12
 [1.3.11]: https://github.com/Automattic/jetpack-error/compare/v1.3.10...v1.3.11
 [1.3.10]: https://github.com/Automattic/jetpack-error/compare/v1.3.9...v1.3.10