+++ projects/packages/status/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.0] - 2022-01-25
+### Added
+- Added Visitor class for status regarding the site visitor.
+
 ## [1.9.5] - 2022-01-04
 ### Changed
 - Switch to pcov for code coverage.
@@ -146,6 +150,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Introduce a status package
 
+[1.10.0]: https://github.com/Automattic/jetpack-status/compare/v1.9.5...v1.10.0
 [1.9.5]: https://github.com/Automattic/jetpack-status/compare/v1.9.4...v1.9.5
 [1.9.4]: https://github.com/Automattic/jetpack-status/compare/v1.9.3...v1.9.4
 [1.9.3]: https://github.com/Automattic/jetpack-status/compare/v1.9.2...v1.9.3